feat(win): reuse filebox & msgbox
