+++ winio-ui-win32/Cargo.toml
@@ -12,7 +12,7 @@ repository = { workspace = true }
 winio-primitive = { workspace = true }
 winio-handle = { workspace = true, features = ["win32"] }
 winio-pollable = { workspace = true }
-winio-ui-windows-common = { workspace = true }
+winio-ui-windows-common = { workspace = true, features = ["win32"] }
 
 compio = { workspace = true, features = ["arrayvec"] }
 compio-log = { workspace = true }
