+++ winio-ui-winui/Cargo.toml
@@ -13,7 +13,7 @@ winio-primitive = { workspace = true }
 winio-handle = { workspace = true, features = ["winui"] }
 winio-callback = { workspace = true }
 winio-pollable = { workspace = true }
-winio-ui-windows-common = { workspace = true }
+winio-ui-windows-common = { workspace = true, features = ["winui"] }
 
 compio = { workspace = true }
 compio-runtime = { workspace = true, features = ["notify-always"] }
